requirements.txt
google-auth==1.20.1
google-auth-oauthlib==0.4.1
gspread==3.6.0
numpy==1.18.4
oauth2client==4.1.3
oauthlib==3.1.0
Pillow==7.1.2
tweepy==3.9.0